requirements.txt
twikit
python-dotenv
requests
typing-extensions
black
pytest